Complete

+++ Tiki_Shop/src/main/java/com/qnit18/Tiki_Shop/repo/AttributeRepository.java
@@ -3,5 +3,5 @@ package com.qnit18.Tiki_Shop.repo;
 import com.qnit18.Tiki_Shop.entity.Attribute;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttributeRepository extends JpaRepository<Attribute, Integer> {
+public interface AttributeRepository extends JpaRepository<Attribute, Long> {
 }
