+++ Tiki_Shop/src/main/java/com/qnit18/Tiki_Shop/repo/QuantitySoldRepository.java
@@ -3,5 +3,5 @@ package com.qnit18.Tiki_Shop.repo;
 import com.qnit18.Tiki_Shop.entity.QuantitySold;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuantitySoldRepository extends JpaRepository<QuantitySold, Integer> {
+public interface QuantitySoldRepository extends JpaRepository<QuantitySold, Long> {
 }
