+++ Tiki_Shop/src/main/java/com/qnit18/Tiki_Shop/repo/SpecificationRepository.java
@@ -3,5 +3,5 @@ package com.qnit18.Tiki_Shop.repo;
 import com.qnit18.Tiki_Shop.entity.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecificationRepository extends JpaRepository<Specification, Integer> {
+public interface SpecificationRepository extends JpaRepository<Specification, Long> {
 }
